Implement three design/feature improvements to CodeKata
1. Redesign Invitations Page with Modern Tabbed Interface
   - New tabbed navigation showing Received, Sent, Friends, and Find People
   - Modern card-based design with user avatars and status badges
   - Better visual hierarchy with gradient avatars
   - Quick navigation between tabs
   - Improved user experience with inline actions

2. Redesign User Profile with Modern Layout
   - Gradient header with user initials avatar
   - Three stat cards showing challenges solved, duels won, and total duels
   - Tabbed content area for questions and answers
   - Sidebar with action buttons and friends list
   - Responsive grid layout for desktop and mobile
   - Better visual organization with clearer typography

3. Restrict Categories to Admin Only
   - Admin categories at /admin/categories/* with full CRUD
   - Public categories at /categories/* with index/show only
   - Categories controller now only handles view operations
   - New Admin::CategoriesController with BaseController inheritance
   - Routes properly namespace admin resources
   - Non-admin users cannot create, edit, or delete categories

All features tested and working correctly.
